fix: never render HeaderBar without runtime provider (#587)
* fix: never render HeaderBar without runtime provider

* style: formatting

* fix: log locale initialization in useEffect, add emoji to log

* fix: short-circuit the system info fetch if no url provided

* style: sort propTypes
