+++ adapter/src/components/AppWrapper.js
@@ -1,0 +1,29 @@
+import { HeaderBar } from '@dhis2/ui'
+import PropTypes from 'prop-types'
+import React from 'react'
+import { useCurrentUserLocale } from '../utils/useLocale.js'
+import { Alerts } from './Alerts.js'
+import { LoadingMask } from './LoadingMask.js'
+import { styles } from './styles/AppWrapper.style.js'
+
+export const AppWrapper = ({ appName, children }) => {
+    const { loading } = useCurrentUserLocale()
+
+    if (loading) {
+        return <LoadingMask />
+    }
+
+    return (
+        <div className="app-shell-adapter">
+            <style jsx>{styles}</style>
+            <HeaderBar appName={appName} />
+            <div className="app-shell-app">{children}</div>
+            <Alerts />
+        </div>
+    )
+}
+
+AppWrapper.propTypes = {
+    appName: PropTypes.string.isRequired,
+    children: PropTypes.node,
+}
