+++ adapter/src/components/ServerVersionProvider.js
@@ -1,9 +1,10 @@
 import { Provider } from '@dhis2/app-runtime'
 import PropTypes from 'prop-types'
 import React, { useEffect, useState } from 'react'
-import { get } from '../utils/api'
-import { parseServerVersion } from '../utils/parseServerVersion'
-import { LoadingMask } from './LoadingMask'
+import { get } from '../utils/api.js'
+import { parseServerVersion } from '../utils/parseServerVersion.js'
+import { LoadingMask } from './LoadingMask.js'
+import { LoginModal } from './LoginModal.js'
 
 export const ServerVersionProvider = ({ url, apiVersion, children }) => {
     const [{ loading, error, systemInfo }, setState] = useState({
@@ -11,6 +12,11 @@ export const ServerVersionProvider = ({ url, apiVersion, children }) => {
     })
 
     useEffect(() => {
+        if (!url) {
+            setState({ loading: false, error: new Error('No url specified') })
+            return
+        }
+
         setState(state => (state.loading ? state : { loading: true }))
         const request = get(`${url}/api/system/info`)
         request
@@ -31,7 +37,7 @@ export const ServerVersionProvider = ({ url, apiVersion, children }) => {
     }
 
     if (error) {
-        return children
+        return <LoginModal />
     }
 
     const serverVersion = parseServerVersion(systemInfo.version)
@@ -52,7 +58,7 @@ export const ServerVersionProvider = ({ url, apiVersion, children }) => {
 }
 
 ServerVersionProvider.propTypes = {
-    url: PropTypes.string.isRequired,
     apiVersion: PropTypes.number,
     children: PropTypes.element,
+    url: PropTypes.string,
 }
